LEAF refactor
