+++ speechbrain/processing/signal_processing.py
@@ -6,6 +6,7 @@ Authors
  * Francois Grondin 2020
  * William Aris 2020
  * Samuele Cornell 2020
+ * Sarthak Yadav 2022
 """
 import torch
 import math
@@ -528,3 +529,92 @@ def resynthesize(enhanced_mag, noisy_inputs, stft, istft, normalize_wavs=True):
         pred_wavs = normalize(pred_wavs, amp_type="peak")
 
     return pred_wavs
+
+
+def gabor_impulse_response(t, center, fwhm):
+    """
+    Function for generating gabor impulse responses
+    as used by GaborConv1d proposed in
+
+    Neil Zeghidour, Olivier Teboul, F{\'e}lix de Chaumont Quitry & Marco Tagliasacchi, "LEAF: A LEARNABLE FRONTEND
+    FOR AUDIO CLASSIFICATION", in Proc of ICLR 2021 (https://arxiv.org/abs/2101.08596)
+    """
+    denominator = 1.0 / (torch.sqrt(torch.tensor(2.0) * math.pi) * fwhm)
+    gaussian = torch.exp(
+        torch.tensordot(
+            1.0 / (2.0 * fwhm.unsqueeze(1) ** 2),
+            (-(t ** 2.0)).unsqueeze(0),
+            dims=1,
+        )
+    )
+    center_frequency_complex = center.type(torch.complex64)
+    t_complex = t.type(torch.complex64)
+    sinusoid = torch.exp(
+        torch.complex(torch.tensor(0.0), torch.tensor(1.0))
+        * torch.tensordot(
+            center_frequency_complex.unsqueeze(1),
+            t_complex.unsqueeze(0),
+            dims=1,
+        )
+    )
+    denominator = denominator.type(torch.complex64).unsqueeze(1)
+    gaussian = gaussian.type(torch.complex64)
+    return denominator * sinusoid * gaussian
+
+
+def gabor_impulse_response_legacy_complex(t, center, fwhm):
+    """
+    Function for generating gabor impulse responses, but without using complex64 dtype
+    as used by GaborConv1d proposed in
+
+    Neil Zeghidour, Olivier Teboul, F{\'e}lix de Chaumont Quitry & Marco Tagliasacchi, "LEAF: A LEARNABLE FRONTEND
+    FOR AUDIO CLASSIFICATION", in Proc of ICLR 2021 (https://arxiv.org/abs/2101.08596)
+    """
+    denominator = 1.0 / (torch.sqrt(torch.tensor(2.0) * math.pi) * fwhm)
+    gaussian = torch.exp(
+        torch.tensordot(
+            1.0 / (2.0 * fwhm.unsqueeze(1) ** 2),
+            (-(t ** 2.0)).unsqueeze(0),
+            dims=1,
+        )
+    )
+    temp = torch.tensordot(center.unsqueeze(1), t.unsqueeze(0), dims=1)
+    temp2 = torch.zeros(*temp.shape + (2,), device=temp.device)
+
+    # since output of torch.tensordot(..) is multiplied by 0+j
+    # output can simply be written as flipping real component of torch.tensordot(..) to the imag component
+
+    temp2[:, :, 0] *= -1 * temp2[:, :, 0]
+    temp2[:, :, 1] = temp[:, :]
+
+    # exponent of complex number c is
+    # o.real = exp(c.real) * cos(c.imag)
+    # o.imag = exp(c.real) * sin(c.imag)
+
+    sinusoid = torch.zeros_like(temp2, device=temp.device)
+    sinusoid[:, :, 0] = torch.exp(temp2[:, :, 0]) * torch.cos(temp2[:, :, 1])
+    sinusoid[:, :, 1] = torch.exp(temp2[:, :, 0]) * torch.sin(temp2[:, :, 1])
+
+    # multiplication of two complex numbers c1 and c2 -> out:
+    # out.real = c1.real * c2.real - c1.imag * c2.imag
+    # out.imag = c1.real * c2.imag + c1.imag * c2.real
+
+    denominator_sinusoid = torch.zeros(*temp.shape + (2,), device=temp.device)
+
+    denominator_sinusoid[:, :, 0] = (
+        denominator.view(-1, 1) * sinusoid[:, :, 0]
+    ) - (torch.zeros_like(denominator).view(-1, 1) * sinusoid[:, :, 1])
+
+    denominator_sinusoid[:, :, 1] = (
+        denominator.view(-1, 1) * sinusoid[:, :, 1]
+    ) + (torch.zeros_like(denominator).view(-1, 1) * sinusoid[:, :, 0])
+
+    output = torch.zeros(*temp.shape + (2,), device=temp.device)
+
+    output[:, :, 0] = (denominator_sinusoid[:, :, 0] * gaussian) - (
+        denominator_sinusoid[:, :, 1] * torch.zeros_like(gaussian)
+    )
+    output[:, :, 1] = (
+        denominator_sinusoid[:, :, 0] * torch.zeros_like(gaussian)
+    ) + (denominator_sinusoid[:, :, 1] * gaussian)
+    return output
